requirements.txt
streamlit==1.22.0
opencv-python==4.7.0.72
ultralytics==8.0.110
pillow==9.5.0
numpy==1.24.3
ultralytics